requirements.txt
flask==3.1.2
flask-cors==6.0.1
psycopg2-binary==2.9.10
